ST-27-refactor: 프로필에디터 위치 변경 및 코드 수정
